README: update filenames and locations
The main README points at conf/autobuilder.conf and conf/yoctoAB.conf,
both of which aren't the current places where the autobuilder looks for
these files anymore.

README-NEW-AUTOBUILDER has a bit of whitespace damage, in the text and in
the examples, so clean that up to make it more paste-friendly.

Signed-off-by: Joe MacDonald <joe_macdonald@mentor.com>