gitignore: add janitor log and additional config files
The janitor log for workers was missing from gitignore, so add it.

Every deployment is expected to make their own changes to buildset-config,
so add the whole directory to the ignore list.

Similarly, every deployment is likely to make their own changes to
config/autobuilder.conf, so ignore changes to it as well.

Signed-off-by: Joe MacDonald <joe_macdonald@mentor.com>