CheckBSPExists.py: smart way of skipping build steps
This commit does a few things:

First, we can't do an integer compare on layerversion_intel
as it's a str, so we need to convert all that.

Second, let's be smart and skip the rest of the build if we
don't need to build it:

                self.finished(SUCCESS)
                self.build.allStepsDone()

will set the step to SUCCESS and then tell the build that
we just ran the last step (even though we didn't).

Signed-off-by: Beth Flanagan <elizabeth.flanagan@intel.com>