PublishArtifacts: enable publishing of sstate
The PUBLISH_SSTATE variable in config/autobuilder.conf didnt' appear to be
wired up anywhere so I've forced it into PublishArtifacts here.

I consider this patch an RFC because a) it required both the PUBLISH_SSTATE
variable to be set and b) an sstate parameter to be passed to the
PublishArtifacts step. Having to turn things on twice is "not good".

Signed-off-by: Joshua Lock <joshua.lock@collabora.co.uk>