AddKernelProps: Split KBRANCH/KMETA and construct SRC_URI
[YOCTO #7685]
[YOCTO #7688]

This commit does two things.

First, KMETA and KBRANCH can have different SRC_URIs. This allows us to set them
independently of each other.

This also greatly simplifies the user input for SRC_URI for linux-yocto. All an
end user needs to input is:

a git URI for either/or both KMETA/KBRANCH,
a git branch for each (optional), defaults to KMETA/KBRANCH in the recipe
a git hash for each (optional), defaults to AUTOREV

This should be used with the next commit of buildset-config.kernel.

Signed-off-by: Beth Flanagan <elizabeth.flanagan@intel.com>