buildset-config.kernel: Add override functionality
This commit adds the ability to override KMETA/KBRANCH/SRC_URI
for the kernel and allows the nightly build to pass those
properties to the triggered builds.

This buildset, btw, is a good example of the set_properties
property of the TriggerBuilds, in how to pass custom props to
other builds.

Signed-off-by: Beth Flanagan <elizabeth.flanagan@intel.com>