CheckOutToasterLayers.py: Use a simple git fetcher for toaster
For toaster builds because we have no idea what we'll be fetching
prior to build time we need a bit different of a git fetcher.

This is a simple one which takes all the layers passed in from
a toaster POST string and then puts them together into
something reasonable.

Signed-off-by: Beth Flanagan <elizabeth.flanagan@intel.com>