yoctogit.py: Fix error handling
Currently, if several of the commands in _fetch() fail, the build merrily
continues and ignores the errors. This effectively means that some random
hash is built but is reported as the hash we were trying to build. In
other words, we can't trust what the autobuilder is doing :(.

This patch improves the error handling so that if an error occurs in any
of the commands, it propagates back up the chain. The retry code can then
handle as appropriate.

This code was tested by changing each command in turn to something invalid
and ensuring the fetch buildstep failed in each case.

The underlying trigger for this issue was a branch head that was force
pushed during a build. That issue still remains but at least it will
error now rather than just building something random.

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>