SendErrorReport: Fix for newer layer versions
We bumped to layer version "6" and this code broke resulting in:

'/bin/sh: -c: line 1: syntax error: unexpected end of file' on the autobuilder.
This was due to an incomplete shell command.

This commit fixes up the conditional for layer verison 5 and later and tidies up
a few of the error paths too.

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>