buildset-config.controller: don't specify test suites explicitly
The autobuilder was specifying the test list explicitly in several buildsets
because in the past the SMART and RPM tests didn't skip automatically on non-RPM
images.  However we're removing the dmesg test as it is obsolete (replaced by
parselogs) but the autobuilder still attempts to run it.

Now that the RPM and SMART tests skip correctly, remove all explict test suite
lists and instead either let the image specify what test suites to run, or use
suitesappend.

Signed-off-by: Ross Burton <ross.burton@intel.com>