buildset-config.controller: Workaround module dependency issue
Dynamically loaded modules are a problem for multilib builds: 32-bit
Pango and Gdk-Pixbuf modules in particular do not get included in the
image when 32-bit versions of the libraries get included.

Proper fix may require extensive changes to dependency handling:
Include the required modules manually in the test configuration as a
workaround for 2.0.

Fixes [YOCTO #8624].
The underlying issue is [YOCTO #8570]

Signed-off-by: Jussi Kukkonen <jussi.kukkonen@intel.com>