CreateBBLayersConf.py: Refactor layer logic for qt3/4
In order to maintain backwards compatibility for the qt3/qt4 layers
we need to check LAYERVERSION_core and apply the appropriate layer.

That said, some of the logic around how we were doing layers was faulty.
We now ensure that entries into BBLAYERS shouldn't duplicate. I've also
cleaned up some of the code that could have made it so that a layer,
even though it wasn't defined, wouldn't have been added.

Signed-off-by: Beth Flanagan <elizabeth.flanagan@intel.com>