/bin/release_scripts: Release automation scripts
[YOCTO #8942]

This is the initial commit of the release.py and misc component scripts which
form the basis for release automation work. The release.py script may be used
to stage and publish a major, point, or milestone release. The various
standalone scriptlet perform individual steps of parts of the release
 process and are handy for clean up, incremental staging, and testing.

Signed-off-by: Graydon, Tracy <tracy.graydon@intel.com>