GetLayerVersion.py: Fix how we find layerversion.
Because we've got LAYERVERSION_yocto living in meta-poky, we need
to switch how we do this. When we made initial changes here, we
had assumed the meta-poky layer would have a LAYERVERSION_poky.

We can't change this in config files because then we'd end up with
two dict entries named meta-yocto which would mess things up.

Signed-off-by: Elizabeth Flanagan <elizabeth.flanagan@intel.com>