bin/release_scripts/release.py: add a check for valid rc candidate
This is part of the release automation work, related to Yocto bug #8942. This
patch adds a check for a valid release candidate ID passed in from command line
and quits if it is bogus. Adds a few comments/output for the user for the
various release steps to improve clarity.

Signed-off-by: Graydon, Tracy <tracy.graydon@intel.com>