bin/release_scripts/release.py: add ADT publishing feature to release script
This is part of the release automation work, related to Yocto bug #8942. This
patch adds the ADT repo publishing functionality for backward compatability
for point releases, etc. Not used going forward since we have moved to the
eSDK instead, so this defaults to NOT publishing the ADT repo. A flag has
been added to cue publishing, along with an option to provide a custom
target direactory name for oddball cases where release numbering may not
be linear. i.e There was no 1.9 release, so ADT repo for 2.0_M1 would have
actually been 1.8+snapshot. This allows for those corner cases.

Signed-off-by: Graydon, Tracy <tracy.graydon@intel.com>