RunESDKSanityTests.py: Adding ESDK sanity and fixing builds
eSDK sanity testing was being done incorrectly. This adds an additional
build that runs it right.

This commit also fixes the eSDK builds. We need to run and test sato
and minimal for both i686 and x86_64. This adds those builds in.

It also skips those builds if LAYERVERSION_core is < 7.

Signed-off-by: Elizabeth Flanagan <elizabeth.flanagan@intel.com>