PublishArtifacts.py/buildtools.conf: Fix deploy/build issues
uninative build needs to have both SDKMACHINES. Also, we require
a change in the uninative tarball name. The issue seems to be
that when we build both, one stomps the other as the tarball name
is not fine grained enough to differentiate between the two builds.

Once uninative.bbclass is patched so that:
UNINATIVE_TARBALL ?= "${BUILD_ARCH}-nativesdk-libc.tar.bz2"
is
UNINATIVE_TARBALL ?= "${BUILD_ARCH}-nativesdk-libc-${SDKMACHINE}.tar.bz2"

We should have working deploys again.

Signed-off-by: Elizabeth Flanagan <elizabeth.flanagan@intel.com>