ptest-*.conf: Remove ptests, add them to nightly.
This removes the ptest-*.conf builds, moves them into nightly-*.conf
and then adds ptests for arm/mips/ppc.

As we're moving them into the main nightly* builds, we have to
switch DEPLOY_DIR to deploy-ptest to ensure we get the correct
images. PublishArtifacts was modified to ensure we pick them up.

Signed-off-by: Elizabeth Flanagan <elizabeth.flanagan@intel.com>