BuildImages.py: Backwards compatability skipping of Images
We've some images/features that are unsupported in earlier
releases. This commit checks if those do not exist in
the LAYERVERSION_core and if so skips.

This commit also cleans up some of that logic and improves
on how we do that skipping by removing the echo command and
just issues a self.finished(SUCCESS)

Signed-off-by: Elizabeth Flanagan <elizabeth.flanagan@intel.com>