ab-prserv: Autobuilder PRSERV support
[YOCTO #7690]

This adds support for the PR service to the autobuilder.

It introduces two new autobuilder.conf fields:

PRSERV_HOST: the pr service host name or ip. default is localhost
PRSERV_PORT: the port used for the pr service. default is 0

It adds a git clone/git pull to starting and stopping so that
bitbake-prserv is up to date on every restart.

We only add the needed bits to auto.conf if all the logic for
buildhistory is also in place.

Signed-off-by: Elizabeth Flanagan <elizabeth.flanagan@intel.com>