CreateAutoConf.py: Add multibranch/repo buildhistory support
This commit adds multibranch/multirepo buildhistory support.
You will need to restructure your buildhistory repo in order to
keep your current history. It stores history in a <short_repo_name>/
<branch>/<buildername> branch format.

It also introduces the concept of "throwaway" history. Example, I
want to compare master-next and master but as I can be assured that
master-next won't be rebased, on every build of it, I remove the remote
branch and reinitialize it.

Signed-off-by: Elizabeth Flanagan <elizabeth.flanagan@intel.com>