bin/buildlogger: improve compatibility with older versions of requests
The json() convenience method isn't available in older versions of the
requests module, such as the 0.8.2 version available in the repositories
for Ubuntu 12.04 LTS. Improve compatibility with older requests versions by
using our parse_json() method always, instead of only when talking to the
wiki.

Add some additional logic to parse_json() to paper over API differences
between requests 0.8.2 (Ubuntu 12.04) and requests 2.1.10 (Fedora 23).

Furthermore print an error when we fail to import the requests module, as
it's not available in a standard python install.

Signed-off-by: Joshua Lock <joshua.g.lock@intel.com>