CreateAutoConf.py: Add optional path and rm of auto.conf
Normally, we assume auto.conf is in ./build/conf/auto.conf
But some people have different layer setups. This allows us
to set the path of the auto.conf with a default to it's normal
path.

This commit also allows us to not remove the auto.conf. This
was requested from end users because sometimes they're using
a script generated auto.conf that CreateAutoConf ends up
stomping.

Signed-off-by: Beth 'pidge' Flanagan <pidge@toganlabs.com>