PublishArtifacts.py: create md5sum files in same dir as artifacts
Previously all the md5sum files got put into the top level deploy directory.
This patch keeps the md5sum file in the same directory as the file it is
hashing. It also limits the traversal depth to five, to avoid hashing the
wic components that go into making the wic images.

Signed-off-by: Bill Randle <william.c.randle@intel.com>