BuildImages.py: build toaster eventlog only when requested
Qualify creation of the toaster eventlog based on the setting of the new
'custom_create_eventlog' property. Place the eventlog in build/tmp/log,
instead of build and add a timestamp to the eventlog filename since
multiple event logs can be generated during a multi-image buildset and
otherwise only the last one would otherwise be saved.

[YOCTO #9884]

Signed-off-by: Bill Randle <william.c.randle@intel.com>