BuildSet.py: allow using local git mirror even if OPTIMIZED_GIT_CLONE is False
OPTIMIZED_GIT_CLONE controls the use of two features: a local git mirror
directory and an optimized move/copy of old build directories. Allow the
use of the local mirror, even if the optimized move/copy is not used.

To disable the local mirror, comment out the setting of OGIT_MIRROR_DIR,
as well as setting OPTIMIZED_GIT_CLONE False.

[YOCOT #10010]

Signed-off-by: Bill Randle <william.c.randle@intel.com>