buildlogger: fix sorting of recent builds
Recent builds, >999, on the YP AutoBuilder were not being logged.
This is because we sort the builds based on a string representation
of their build number, yet string 1001 < string 999.

Resolve this by converting the build id's to an int, sorting them
and then determining which is the highest/most recent build id.

Signed-off-by: Joshua Lock <joshua.g.lock@intel.com>