CreateAutoConf: fix check for remote buildhistory branch
The [--exit-code] option to git ls-remote must precede the <repository>
and <refs> arguments, otherwise it isn't recognised and git ls-remote
always returns 0 regardless of whether the specified ref is found in
the repository.

$ git ls-remote git://git.yoctoproject.org/poky-buildhistory refs/heads/master-next --exit-code
$ echo $?
0

$ git ls-remote --exit-code git://git.yoctoproject.org/poky-buildhistory refs/heads/master-next
$ echo $?
2

Failing to correctly determine whether a branch exists at the remote
results in the local branch not being created and ultimately in
failures to push buildhistory to the remote.

Signed-off-by: Joshua Lock <joshua.g.lock@intel.com>