CreateAutoConf: finish fix of check for remote buildhistory branch
The fix in 61b4bf4b463b7859ab025755c9c232688503a70f was incomplete
because the exit code was still masked by piping the output of
git ls-remote into a cut. The manipulated output of the git
ls-remote doesn't appear to be used anywhere else during this
buildstep so this change removes the cut and sed.

[YOCTO #10171]

Signed-off-by: Joshua Lock <joshua.g.lock@intel.com>