CreateAutoConf: fix deleting remote throwaway buildhistory branch
83c609eddbc88f07cbdb0d29813132873266d476 moved the deletion of
remote throwaway branches until after the presence of the branch
had been tested for, however it erroneously changed the command so
that the remote branch is deleted *after* an empty one is pushed,
breaking the intended behaviour.

This change makes an extra git ls-remote call for throwaway
branches so that an existing remote branch is deleted *before* the
empty branch is pushed.

We also redirect the output of git ls-remote to /dev/null so that
we don't pollute the stdio logs on the builders.

Signed-off-by: Joshua Lock <joshua.g.lock@intel.com>