PublishArtifacts: include dtb's for qemuarm
The new 4.8 kernel in morty generates dtb files for qemuarm,
however the generated files don't include qemuarm in the filename
and thus aren't picked up by the glob we pass to cp.

Resolve this for now by copying *Image* too when qemuarm is in
artifacts, much as we do for beaglebone.

[YOCTO #10435]

Signed-off-by: Joshua Lock <joshua.g.lock@intel.com>