CreateAutoConf: set a property when this is a release build
We want some build steps (CreateAutoConf and PublishArtifacts) which
run on a triggered build to behave differently for a release build.
Ideally we'd just propage the custom_release_me property to the
triggered builds using copy_properties, however as several
buildsteps are already using this property in conjunction with a
series of other properties on the triggering build we don't want to
propagate too much unrelated (for our needs) state into triggered
builds and thus take the option to set a new property.

Signed-off-by: Joshua Lock <joshua.g.lock@intel.com>