buildsteps: add skipStep() method to BitbakeShellCommand
Unify the logic for skipping a buildstep in a helper method on the
parent BitbakeShellCommand class.

When called skipStep() will set a dummy command (echoing the
given reason) and update the buildstep's description. This helps
us ensure consistent skipping behaviour amongst build steps.

Note: we're purposfully not finising the step early with a call to
self.finished() as calling that method before ShellCommand.start()
in conjunction with the commandCompleted() callback in the parent
BitbakeShellCommand class was throwing exceptions which propagated
errors throughout builders.

Signed-off-by: Joshua Lock <joshua.g.lock@intel.com>