CreateAutoConf: include a toolchain in minimal eSDK
In order to ensure the generated extensible SDK's can still be
used to succesfully run the sanity tests we need to include a
usable toolchain.

This makes the minimal eSDK larger but ensures that the eSDK is
usable from an IDE, or other tools, without having to perform
additional steps.

Signed-off-by: Joshua Lock <joshua.g.lock@intel.com>